changed slider value type from int to long

+++ src/tl/GUIutil/GUIValueFunction.java
@@ -2,5 +2,5 @@ package tl.GUIutil;
 
 public interface GUIValueFunction
 {
-	public void execute(int value, GUIControl control);
+	public void execute(long value, GUIControl control);
 }
